completed basic admin functionality

+++ ABETAccreditationStation/src/main/java/com/maccari/abet/web/controller/UserController.java
@@ -1,6 +1,5 @@
-package com.maccari.abet.web;
+package com.maccari.abet.web.controller;
 
-import java.security.Principal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -17,12 +16,16 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.maccari.abet.domain.entity.User;
+import com.maccari.abet.domain.entity.WebUser;
 import com.maccari.abet.domain.service.UserService;
+import com.maccari.abet.web.UserValidator;
+import com.maccari.abet.web.WebUserValidator;
 
 @Controller
 public class UserController {
@@ -32,26 +35,12 @@ public class UserController {
 	@Autowired
 	private UserValidator userValidator;
 	
+	@Autowired
+	private WebUserValidator webUserValidator;
+	
 	@Lazy
 	@Autowired
 	private PasswordEncoder passwordEncoder;
-	
-	@RequestMapping(value = "/")
-	public String homePage(Principal principal, Model model) {
-		String email = "user email";
-		if(principal != null) {
-			principal.getName();
-			email = principal.getName();
-		}
-		model.addAttribute("msg", email);
-		
-		return "home/index";
-	}
-	
-	@RequestMapping(value = "/edit")
-	public String edit() {
-		return "home/edit";
-	}
 	
 	@RequestMapping(value = "/manage")
 	public String manage(Model model) {
@@ -78,14 +67,9 @@ public class UserController {
 		
 		List<String> roles = new ArrayList<String>();
 		roles.add("ROLE_GENERAL");
-		/*for(String role : user.getRoles()) {
-			roles.add("ROLE_" + role);
-		}*/
 
 		user.setRoles(roles);
 		userService.create(user);
-		
-		//autologin(user, roles);
 		
 		return "redirect:/manage";
 	}
@@ -100,9 +84,39 @@ public class UserController {
 	
 	@RequestMapping(value = "/remove", method = RequestMethod.POST)
 	public String removeUser(@RequestParam(value = "email", required = true) String email, Model model) {
+		if(email.isEmpty()) {
+			return "redirect:/manage";
+		}
 		User user = userService.getUserByEmail(email);
 		userService.remove(user);
 		
+		return "redirect:/manage";
+	}
+	
+	@RequestMapping(value = "/edit")
+	public String editUser(@RequestParam(value = "email", required = true) String email, WebUser webUser,
+			Model model) {
+		User user = userService.getUserByEmail(email);
+		model.addAttribute("webUser", new WebUser(user.getEmail(), user.getRoles()));
+		
+		return "user/edit";
+	}
+	
+	@RequestMapping(value = "/edit", method = RequestMethod.POST, params = "submit")
+	public String editUser(@Valid WebUser webUser, BindingResult bindingResult) {
+		webUserValidator.validate(webUser, bindingResult);
+		if(bindingResult.hasErrors()) {
+			return "user/edit";
+		}
+		
+		System.out.println(webUser.getEmail());
+		userService.update(userService.convertWebUser(webUser));
+		
+		return "redirect:/manage";
+	}
+	
+	@RequestMapping(value = "/edit", method = RequestMethod.POST, params = "cancel")
+	public String cancelUserEdit(@Valid WebUser webUser, BindingResult bindingResult) {
 		return "redirect:/manage";
 	}
 	
@@ -113,5 +127,10 @@ public class UserController {
 		UsernamePasswordAuthenticationToken auth = new UsernamePasswordAuthenticationToken(
 			user.getEmail(), null, authorities);
 		SecurityContextHolder.getContext().setAuthentication(auth);
+	}
+	
+	@ModelAttribute("roleTypes")
+	public ArrayList<String> getRoles(){
+		return (ArrayList<String>) userService.getRoles();
 	}
 }
